[dist] Move es5 fix files into es5/*. Lint the linter. Depend on eslint@~2.2.0 until resolved upstream.
